Reduce number of hints records.
Handling the BOUND action explicitly almost doubles the number of hints
records.  Consequently, it is now handled implicitly by the ADJUST and STEM
actions.

Additionally, hints records are now completely stored as 16bit numbers; this
makes it easier later on to compute the number of arguments and whether any
number is larger than 0xFF (to select PUSHB or PUSHW).