Disable (additional) hinting of composites.
Up to now, ttfautohint was agnostic of composites at all (as is FreeType's
autohinter normally), this is, internal calls to FT_Load_Glyph always
resolved the subglyphs to construct a simple glyph.  This normally has the
advantage that the whole glyph gets hinted at once, possibly improving the
overall shape.  However, using the normal TrueType hinting mechanism, it is
not possible to switch off hinting of single glyph components; as an effect,
each subglyph got hinted, and then the whole glyph again.  If there were no
scaling, this wouldn't do any harm (except being slower), but the autohinter
actually uses scaling...

This commit disables the creation of hints for composite glyphs (but glyph
components get hints).  At the same time, the default hinting module is set
to `latin' so that all glyphs in a font get hinted.  The offsets of glyph
components is still unscaled and thus incorrect (this will be handled by
another patch).

Another benefit is that the size of the output TTF can shrink a lot.  For
pala.ttf, it is a reduction from 493kB to 336kB.