Modify bytecode to create twilight points.
Relying on the Microsoft OpenType specification only was a mistake, since it
lacks the description of how to create twilight points.  [I've reported this
bug, and a future version of the specification should have this fixed
eventually.]

This patch creates twilight points in accordance with Apple's TrueType
specification.  In particular, it doesn't rely on the initialization of
twilight points to (0,0).  Doing so has the nice side effect that the new
bytecode is smaller since it is no longer necessary to maintain two sets of
twilight points: `Creating' a twilight point means to set its original
coordinate, thus instructions like GC_cur and GC_orig both work as expected
in the twilight zone also.