Use only CVT to communicate between `prep' and `fpgm'.
Neither Microsoft's nor Apple's specifications say how long values in the
storage area live, so be on the safe side.  Note, however, that this change
doesn't fix the rendering issues with iOS.  Commit 08027dd5 doesn't do this
either, unfortunately.

And just for the record, I have to correct one remark from 08027dd5: The
Apple specification actually does mention that twilight points are
originally positioned at the origin (in the description of the MIAP
instruction).