Improve handling of composite glyphs (finished).
Finally, the last patch of a series which started with commit 9329c207.
Documentation will follow soon.

The idea is to prepend a dummy composite element (called `.ttfautohint') to
all composite glyphs; its only function is to execute some bytecode,
increasing a composite nesting level counter.  Later on we test this counter
to prevent execution of hints which the autohinter has already applied to
composite elements.

In other words, composite glyphs are now hinted as a whole, too.  While this
increases the size of fonts, it greatly improves the rendering of such
glyphs.

Due to the inserted dummy composite element(s), some point indices change.
The previous patches implemented the necessary functionality in SFNT tables;
this commit combines everything to make it work actually.