Uh again.  Patch `bootstrap'.
The original script always assumes a submodule name `gnulib'.  However, I've
set up ait's not an a directory `.gnulib' with `gnu submodule add'
(following the current gnulib documentation), and this git command always
uses the directory name to deduce the submodule name.

Obviously, it's both a documentation bug of gnulib and a deficiency of
`bootstrap'.  Patching it, however, is currently the easiest solution.