Make console output work on the Windows console.
A Windows GUI program normally doesn't provide stdin, stdout, and stderr.
If it does, it always creates a console.  The application can close this
console immediately after its creation, however, there's always a short
visible flash on the screen, which we want to avoid.

The implemented solution is taken from

  http://code.google.com/p/dualsubsystem/

We now have a dummy `ttfautohint.com' which spawns `ttfautohint.exe', using
pipes to redirect the standard streams.  Since Windows prefers `.com' to
`.exe' if you omit the extension, typing things like `ttfautohint --help' on
the console works as expected.  If ttfautohint gets called from the GUI, it
doesn't matter whether the COM or EXE file is used.