Add GDI ClearType support.
There are two ClearType flavours available on Windows: The older GDI
ClearType, introduced in 2000, and the recent DW ClearType, introduced in
2008.  The main difference is that the older incarnation behaves like a B/W
renderer along the y axis, while the newer version does vertical smoothing
also.

The only possibility to differentiate between GDI and DW ClearType is
testing bit 10 in the GETINFO instruction (with return value in bit 17; this
works for TrueType version >= 38), checking whether sub-pixel positioning is
available.

If GDI ClearType is active, we use a different stem width function which
snaps to integer pixels as much as possible.